+++ integration/src/main.ts
@@ -1,0 +1,6 @@
+import { bootstrapApplication } from '@angular/platform-browser';
+
+import { AppComponent } from './app/app.component';
+import { appBrowserConfig } from './app/app.config.browser';
+
+bootstrapApplication(AppComponent, appBrowserConfig);