+++ integrations/hello-world-ng19/src/app/app.config.ts
@@ -1,8 +1,0 @@
-import { provideExperimentalZonelessChangeDetection } from '@angular/core';
-import { provideRouter } from '@angular/router';
-
-import { provideNgxs } from './store';
-
-export const appConfig = {
-  providers: [provideExperimentalZonelessChangeDetection(), provideRouter([]), provideNgxs()]
-};