+++ packages/form-plugin/tests/issues/issue-1567-ngxs-onchanges-hook-with-plugin.spec.ts
@@ -1,4 +1,4 @@
-import { Component, Injectable } from '@angular/core';
+import { Component, inject, Injectable } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
 import { FormBuilder, ReactiveFormsModule } from '@angular/forms';
 import { NgxsModule, NgxsSimpleChange, State, Store } from '@ngxs/store';
@@ -42,11 +42,9 @@ describe('ngxsOnChanges with form plugin (https://github.com/ngxs/store/issues/1
     standalone: false
   })
   class TestComponent {
-    form = this._fb.group({
+    form = inject(FormBuilder).group({
       name: ['']
     });
-
-    constructor(private _fb: FormBuilder) {}
   }
 
   const testSetup = () => {