+++ packages/storage-plugin/schematics/src/_testing/schematics.ts
@@ -1,8 +1,8 @@
 import { SchematicTestRunner, UnitTestTree } from '@angular-devkit/schematics/testing';
-import { workspaceRoot } from '@nrwl/devkit';
+import { workspaceRoot } from '@nx/devkit';
 import { Schema as ApplicationOptions } from '@schematics/angular/application/schema';
 import { Schema as WorkspaceOptions } from '@schematics/angular/workspace/schema';
-import * as path from 'path';
+import * as path from 'node:path';
 
 const angularSchematicRunner = new SchematicTestRunner(
   '@schematics/angular',