+++ packages/store/src/symbols.ts
@@ -56,7 +56,7 @@ export class NgxsConfig {
    * It makes sense to use it only during development to ensure there're no state mutations.
    * When building for production, the `Object.freeze` will be tree-shaken away.
    */
-  developmentMode: boolean;
+  developmentMode!: boolean;
   compatibility: {
     /**
      * Support a strict Content Security Policy.
@@ -76,7 +76,7 @@ export class NgxsConfig {
   };
 }
 
-export { StateOperator };
+export type { StateOperator };
 
 /**
  * State context provided to the actions in the state.
@@ -137,12 +137,3 @@ export interface NgxsAfterBootstrap {
 }
 
 export type NgxsModuleOptions = Partial<NgxsConfig>;
-
-/** @internal */
-declare global {
-  const ngDevMode: boolean;
-  // Indicates whether the application is operating in server-rendering mode.
-  // `ngServerMode` is a global flag set by Angular CLI.
-  // https://github.com/angular/angular-cli/blob/b4e9a2af9e50e7b65167d0fdbd4012023135e875/packages/angular/build/src/tools/vite/utils.ts#L102
-  const ngServerMode: boolean;
-}