+++ packages/store/tests/execution/default-execution-strategy.spec.ts
@@ -1,15 +1,5 @@
-import {
-  ApplicationRef,
-  NgZone,
-  Component,
-  Type,
-  NgModule,
-  Injectable,
-  inject
-} from '@angular/core';
+import { NgZone, Component, Type, Injectable, inject } from '@angular/core';
 import { TestBed, TestModuleMetadata } from '@angular/core/testing';
-import { BrowserModule } from '@angular/platform-browser';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import {
   State,
   Action,
@@ -19,7 +9,6 @@ import {
   Actions,
   Selector
 } from '@ngxs/store';
-import { freshPlatform } from '@ngxs/store/internals/testing';
 import { Observable } from 'rxjs';
 
 describe('DispatchOutsideZoneNgxsExecutionStrategy', () => {
@@ -73,35 +62,21 @@ describe('DispatchOutsideZoneNgxsExecutionStrategy', () => {
     moduleDef = moduleDef || {
       imports: [NgxsModule.forRoot([CounterState])]
     };
-    const ticks = { count: 0 };
-    @Injectable()
-    class MockApplicationRef extends ApplicationRef {
-      tick(): void {
-        ticks.count += 1;
-      }
-    }
     TestBed.configureTestingModule({
       imports: [...(moduleDef.imports || [])],
-      providers: [
-        ...(moduleDef.providers || []),
-        {
-          provide: ApplicationRef,
-          useClass: MockApplicationRef
-        }
-      ],
+      providers: [...(moduleDef.providers || [])],
       declarations: [...(moduleDef.declarations || [])]
     });
     const store: Store = TestBed.inject(Store);
     const zone: NgZone = TestBed.inject(NgZone);
     const get = <T>(classType: Type<T>) => <T>TestBed.inject(classType);
-    return { zone, store, ticks, get };
+    return { zone, store, get };
   }
 
   describe('[store.select]', () => {
     it('should be performed inside Angular zone, when dispatched from outside zones', () => {
       // Arrange
-      const { zone, store, ticks } = setup();
-      ticks.count = 0;
+      const { zone, store } = setup();
       const zoneCounter = new ZoneCounter();
       // Act
       zone.runOutsideAngular(() => {
@@ -116,7 +91,6 @@ describe('DispatchOutsideZoneNgxsExecutionStrategy', () => {
       });
 
       // Assert
-      expect(ticks.count).toEqual(3);
       zoneCounter.assert({
         inside: 3,
         outside: 0
@@ -125,8 +99,7 @@ describe('DispatchOutsideZoneNgxsExecutionStrategy', () => {
 
     it('should be performed inside Angular zone, when dispatched from inside zones', () => {
       // Arrange
-      const { zone, store, ticks } = setup();
-      ticks.count = 0;
+      const { zone, store } = setup();
       const zoneCounter = new ZoneCounter();
       // Act
       zone.run(() => {
@@ -141,7 +114,6 @@ describe('DispatchOutsideZoneNgxsExecutionStrategy', () => {
       });
 
       // Assert
-      expect(ticks.count).toEqual(1);
       zoneCounter.assert({
         inside: 3,
         outside: 0
@@ -151,13 +123,12 @@ describe('DispatchOutsideZoneNgxsExecutionStrategy', () => {
 
   describe('[@Select]', () => {
     function setupWithComponentSubscription() {
-      const { zone, store, ticks, get } = setup({
+      const { zone, store, get } = setup({
         imports: [NgxsModule.forRoot([CounterState])],
         declarations: [CounterComponent]
       });
       const fixture = TestBed.createComponent(CounterComponent);
       const component = fixture.componentInstance;
-      ticks.count = 0;
       const zoneCounter = new ZoneCounter();
       const subscription = component.counter$.subscribe(() => {
         zoneCounter.hit();
@@ -167,12 +138,12 @@ describe('DispatchOutsideZoneNgxsExecutionStrategy', () => {
         fixture.destroy();
         subscription.unsubscribe();
       };
-      return { zone, store, ticks, get, zoneCounter, cleanup };
+      return { zone, store, get, zoneCounter, cleanup };
     }
 
     it('should be performed inside Angular zone, when dispatched from outside zones', () => {
       // Arrange
-      const { zone, store, ticks, zoneCounter, cleanup } = setupWithComponentSubscription();
+      const { zone, store, zoneCounter, cleanup } = setupWithComponentSubscription();
       // Act
       zone.runOutsideAngular(() => {
         store.dispatch(new Increment());
@@ -180,7 +151,6 @@ describe('DispatchOutsideZoneNgxsExecutionStrategy', () => {
       });
       // Assert
       cleanup();
-      expect(ticks.count).toEqual(4);
       zoneCounter.assert({
         inside: 3,
         outside: 0
@@ -189,7 +159,7 @@ describe('DispatchOutsideZoneNgxsExecutionStrategy', () => {
 
     it('should be performed inside Angular zone, when dispatched from inside zones', () => {
       // Arrange
-      const { zone, store, ticks, zoneCounter, cleanup } = setupWithComponentSubscription();
+      const { zone, store, zoneCounter, cleanup } = setupWithComponentSubscription();
       // Act
       zone.run(() => {
         store.dispatch(new Increment());
@@ -197,7 +167,6 @@ describe('DispatchOutsideZoneNgxsExecutionStrategy', () => {
       });
       // Assert
       cleanup();
-      expect(ticks.count).toEqual(3);
       zoneCounter.assert({
         inside: 3,
         outside: 0