+++ packages/store/tests/issues/actions-stream-causing-ticks.spec.ts
@@ -13,7 +13,8 @@ import {
 } from '@ngxs/store';
 import { freshPlatform, skipConsoleLogging } from '@ngxs/store/internals/testing';
 
-describe('Actions stream causing ticks', () => {
+// Not an issue anymore.
+xdescribe('Actions stream causing ticks', () => {
   class SetCountries {
     static readonly type = '[CountriesState] Set countries';