+++ packages/store/types/tests/dispatch.lint.ts
@@ -24,7 +24,7 @@ describe('[TEST]: Action Types', () => {
       imports: [NgxsModule.forRoot()]
     });
 
-    store = TestBed.get(Store);
+    store = TestBed.inject(Store);
   });
 
   it('should be correct type in action decorator', () => {